refactor SigningExtensions to BuildConfigExtensions and add getApiToken function
